requirements-dev.txt
-r requirements.txt
ruff==0.11.5
pytest==8.3.5
pytest-django==4.11.1
playwright==1.51.0
pytest-playwright==0.7.0
pytest-base-url==2.1.0
coverage==7.5.1